syncparty/server/requirements.txt
Django>=5.2,<6.0
channels==4.0.0
channels-redis==4.2.0
djangorestframework>=3.14
 djangorestframework-simplejwt>=5.2
 django-cors-headers>=4.0
